test.js: update most problems to use jsdom & jquery
jquery allows queries of **computed** values of CSS properties. This
allows students to make reasonable substitutions of many variables,
e.g. substituting `margin` for `margin-left`. Doesn't solve al
ltesting problems but it's a good start.
It would be nice to move away entirely form cheerio, but i'm not there
yet :-(

See new `before` clauss in each test suite for the solution to
successfully loading a full working DOM for each web page. That took a
while and a lot of googling to figure out.